chore(aggregated-assets): move the page helper under helpers/
features/platform/README.md prescribes components/ hooks/ helpers/ with only
files that fit no subdirectory at the root. mergeAssetsDataPages is a
cross-feature domain-aware helper, which is what helpers/ is for.
