+++ features/platform/aggregated-assets/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./pagination";
+export * from "./helpers/pagination";
